requirements.txt
# Tensorflow for GAST RTX 3090 (for rcnn)
tensorflow=2.11.1
numpy=1.24.3
psutil=5.9.5
bitarray=2.7.6
scipy=1.11.1
# For the network plots
pydot
graphviz
# Pytorch for GAST RTX 3090 (for GPT-2)
pytorch
torchvision
torchaudio
pytorch-cuda=11.8
transformers=4.29.2
# For data analysis notebooks
ipykernel=6.21.3
matplotlib=3.7.1
pandas=2.0.1